# api.yml
...
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
servers:
  - url: http://localhost
paths:
...
  /get-profile:
    get:
      summary: Get User Profile API
      operationId: get-profile
      parameters:
        - name: "Authorization"
          in: "header"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success Get User Profile
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/GetProfileResponse"
        '403':
          description: Authorization Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    GetProfileResponse:
      type: object
      required:
        - full_name
        - phone_number
      properties:
        full_name:
          type: string
        phone_number:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string